+++ apps/annotation-manager/js/App.tsx
@@ -4,7 +4,10 @@ import {Datastore} from '../../../web/js/datastore/Datastore';
 import {PersistenceLayer} from '../../../web/js/datastore/PersistenceLayer';
 import {Datastores} from '../../../web/js/datastore/Datastores';
 import {DocMeta} from '../../../web/js/metadata/DocMeta';
+import {Text} from '../../../web/js/metadata/Text';
 import {Screenshot} from '../../../web/js/metadata/Screenshot';
+import {AnnotationType} from '../../../web/js/metadata/AnnotationType';
+import {Elements} from '../../../web/js/util/Elements';
 
 const log = Logger.create();
 
@@ -18,12 +21,9 @@ class App<P> extends React.Component<{}, IAppState> {
 
         this.state = {
             annotations: [
-                {
-                    text: 'this is just some text'
-                },
-                {
-                    text: 'this is just some text2'
-                }
+                // {
+                //     text: 'this is just some text2'
+                // }
             ]
         };
 
@@ -32,6 +32,24 @@ class App<P> extends React.Component<{}, IAppState> {
             await this.init();
 
             //this.setState(this.state);
+
+            let fingerprint = '110dd61fd57444010b1ab5ff38782f0f';
+
+            let docMeta = await this.persistenceLayer!.getDocMeta(fingerprint);
+
+            if(docMeta) {
+                let annotations = await this.loadDocMeta(docMeta);
+
+                console.log("FIXME: ", annotations);
+
+                this.setState({
+                    annotations
+                });
+
+            } else {
+                log.error("No docMeta");
+            }
+
 
         })().catch(err => log.error("Could not load disk store: ", err));
 
@@ -50,7 +68,79 @@ class App<P> extends React.Component<{}, IAppState> {
 
     }
 
-    private async loadDocMeta(docMeta: DocMeta) {
+    private async loadDocMeta(docMeta: DocMeta): Promise<IAnnotation[]> {
+
+        let result: IAnnotation[] = [];
+
+        log.info("Loading docMeta...")
+
+        Object.values(docMeta.pageMetas).forEach(pageMeta => {
+
+            Object.values(pageMeta.textHighlights).forEach(textHighlight => {
+
+                console.log("FIXME: ", textHighlight);
+
+                let html: string = "";
+
+                if(typeof textHighlight.text === 'string') {
+                    html = `<p>${textHighlight.text}</p>`
+                }
+
+                if(textHighlight.text instanceof Text) {
+
+                    if(textHighlight.text.TEXT) {
+                        html = `<p>${textHighlight.text.TEXT}</p>`
+                    }
+
+                    if(textHighlight.text.HTML) {
+                        html = textHighlight.text.HTML;
+                    }
+
+                }
+
+                result.push({
+                    annotationType: AnnotationType.TEXT_HIGHLIGHT,
+                    html
+                })
+
+            });
+
+        });
+
+        return result;
+
+    }
+
+    createHTML(annotations: IAnnotation[]) {
+
+        // https://blog.cloudboost.io/for-loops-in-react-render-no-you-didnt-6c9f4aa73778
+        //
+        // let result = []
+        //
+        // annotations.map(annotation => {
+        //     result.push(<div>hello world</div>);
+        // });
+        //
+        // return result;
+
+        {/*{annotations.map((annotation, idx) =>*/}
+        {/*<div className="annotation">*/}
+        {/*{Elements.createElementHTML(annotation.html)}*/}
+        {/*</div>*/}
+        {/*)}*/}
+
+
+        // // Outer loop to create parent
+        // for (let i = 0; i < 3; i++) {
+        //     let children = []
+        //     //Inner loop to create children
+        //     for (let j = 0; j < 5; j++) {
+        //         children.push(<td>{`Column ${j + 1}`}</td>)
+        //     }
+        //     //Create the parent and add the children
+        //     result.push(<tr>{children}</tr>)
+        // }
+        // return result
 
     }
 
@@ -60,11 +150,13 @@ class App<P> extends React.Component<{}, IAppState> {
 
             <div id="annotation-manager">
 
-                {annotations.map((annotation, idx) =>
-                    <div className="annotation">
-                        <div>this is an annotation: {annotation.text}</div>
-                    </div>
-                )}
+                {/*{annotations.map((annotation, idx) =>*/}
+                    {/*<div className="annotation">*/}
+                        {/*{Elements.createElementHTML(annotation.html)}*/}
+                    {/*</div>*/}
+                {/*)}*/}
+
+                {/*{annotations.map((annotation, idx) => Elements.createElementHTML(annotation.html))}*/}
 
             </div>
 
@@ -76,11 +168,11 @@ export default App;
 
 interface IAppState {
 
-
     annotations: IAnnotation[];
 }
 
 interface IAnnotation {
-    text: string
+    annotationType: AnnotationType,
+    html: string
     screenshot?: Screenshot;
 }