+++ web/js/metadata/ExtendedAnnotation.ts
@@ -1,17 +1,19 @@
 import {Annotation} from './Annotation';
 import {Flashcard} from './Flashcard';
+import {Note} from './Note';
+import {Question} from './Question';
 
 export class ExtendedAnnotation extends Annotation {
 
     /**
      * The note for this annotation.
      */
-    public notes: {[key: string]: any} = {};
+    public notes: {[key: string]: Note} = {};
 
     /**
      * The note for this annotation.
      */
-    public questions: {[key: string]: any} = {};
+    public questions: {[key: string]: Question} = {};
 
     /**
      * The note for this annotation.